refactor: moved task execution to util and removed run()
